backend/requirements.txt
Django>=4.0.0
djangorestframework>=3.13.0
psycopg2-binary>=2.9.3
django-cors-headers>=3.10.0
Pillow>=9.0.0
django-filter>=21.1
django-imagekit>=4.1.0
django-extensions>=3.1.5
torch>=2.0.0
numpy>=1.21.0
scikit-learn>=1.0.0
pytz>=2023.3